+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Changelog for TYPO3 CMS Extension pdfviewhelpers
 
-## 1.4.1 - Not yet released
+## 1.5.0 - Not yet released
 - Updates documentation
-- Adds automatic hyphenation to HtmlViewHelper, [#44](https://github.com/bithost-gmbh/pdfviewhelpers/issues/44)
+- Adds automatic hyphenation to all textual ViewHelpers, [#44](https://github.com/bithost-gmbh/pdfviewhelpers/issues/44)
+- Adds support for absolute, relative and TYPO3 EXT: paths, [#43](https://github.com/bithost-gmbh/pdfviewhelpers/issues/43)
 
 ## 1.4.0 - April 10, 2018
 - Fixes posX and posY not working, [#37](https://github.com/bithost-gmbh/pdfviewhelpers/issues/37) (Thanks [@PeterSchuhmann](https://github.com/PeterSchuhmann))